fix: resolve do not set default time
